Expanded functionalities and improved efficiency in ASM package
The commit introduced methods to filter and find static and instance methods and fields of a class. It also created methods that allow search for parent and child methods that match a given method signature. The AsmClassWriter was renamed and optimized, introducing a caching mechanism for retrieved classes. This change also includes the creation of AsmSpec and TestClasses for testing. Tests were updated to utilize the new AsmSpec class.