Implement Identifier Renaming for Deobfuscator
Added a renaming function to the deobfuscator that can rename classes, methods, and fields which have obfuscated names. Also made small changes to ensure empty sets are returned for parents and children methods in the ClassHierarchy when no nodes are present. Made necessary class pool adjustments to accomodate these changes.