Refactored class hierarchy and improved deobfuscator transformations
A significant refactor of the class hierarchy system was undertaken, resulting in the removal of MethodHierarchy and a rewrite of ClassHierarchy as a generic Hierarchy class. The NameMappings class and other relevant classes have been updated accordingly. Additionally, the deobfuscator transformations were enhanced to more effectively rename methods and fields, and a new Transformer subclass VisibilityFixer was introduced. Some changes were also made in the Deobfuscator class to register the VisibilityFixer transformer.