Remove unused Expression Tree (ExprTree) and related classes and methods
Deleted the Expression Tree (ExprTree) class along with related classes which includes Expr, ConstantExpr, NumberExpr, Stmt and TryCatch. Also removed corresponding methods in the MethodNode class related to ExprTree. This was done because these classes and methods were obsolete and were no longer in use in the application. The deletion helps to reduce clutter and improve code readability and maintenance.