Remove ExprOrderFixer and replace it with ExprOrderNormalizer, update multiplier logic and extend ASM classes
The ExprOrderFixer was removed and replaced with a new class called ExprOrderNormalizer. This new class improves the handling of expression order in the code. Furthermore, a significant change was made to adjust the multiplier computation logic to better handle field multiplier encoders and decoders in the MultiplierRemover class.

Additional lines of code were added to handle try catch blocks in the MethodNode extension, and the const expression in ExprTree got expanded to also include ACONST_NULL opcodes. Moreover, new extension properties for FieldNode, ClassNode, and MethodNode were added, namely `origOwner`, `origName`, `origDesc` and `multiplier` for FieldNode; `origName` for ClassNode; and `origOwner`, `origName`, `origDesc`, `opaqueValue` and `opaqueDesc` for MethodNode. These will be used to store original data stripped during the deobfuscation process. Lastly, an initDeob method placeholder was added in a ClassPool extension.